control-table still missing button for unreaded

+++ backend/msgs/utils.py
@@ -2,6 +2,8 @@
 from msgs.models import Msgs
 from users.models import Users 
 from management.utils import get_superior_units
+from draft_forms.models import DraftFormsTable
+from forms.models import FormsTable
 
 
 def new_event_msgs(reference, reporterUnit):
@@ -17,7 +19,9 @@ def delete_event_messages(reference):
     '''
         Delete specific event form msgs row in the database 
     '''
-    if Msgs.objects.filter(reference=reference).exists():
+    # The second condition means xor - if the event appears only in one of the DB, 
+    # It means it doesn't exist in the second db, and therefore the event msgs can be deleted
+    if Msgs.objects.filter(reference=reference).exists() and (FormsTable.objects.filter(reference=reference).exists() ^ DraftFormsTable.objects.filter(reference=reference).exists()):
         event_form_msgs = Msgs.objects.get(reference=reference)
         event_form_msgs.delete()
 
@@ -34,10 +38,29 @@ def update_message_for_relevant_users(sender_unit, reporterUnit, reference):
         user.unreadedMessages[str(reference)] = user.unreadedMessages.get(str(reference), 0) + 1
         user.save()
 
+
 def update_user_read_msg(personal_number, reference):
     try:
         user = Users.objects.get(personalNumber=personal_number)
-        del user.unreadedMessages[str(reference)]
+        if str(reference) in user.unreadedMessages:
+            del user.unreadedMessages[str(reference)]
+            user.save()
         return True
     except:
         return False
+
+
+def update_user_event_deleted(reference, reporterUnit):
+    '''
+        Update all users to delete this event reference from their unreaded msgs
+    '''
+    # If the event appaers in the two DB, it means it shouldn't been deleted - the users still needs to by notified about 
+    # its ureaded msgs
+    if (FormsTable.objects.filter(reference=reference).exists() and DraftFormsTable.objects.filter(reference=reference).exists()):
+        return False
+
+    users_to_delete_event = Users.objects.filter(unit__in=get_superior_units(reporterUnit))
+    for user in users_to_delete_event:
+        if str(reference) in user.unreadedMessages:
+            del user.unreadedMessages[str(reference)]
+            user.save()
